Refactor trybuild test support
